Rename binary and python package
